Extract AsyncNotification from model_soup
